backend/data_ingestion/requirements.txt
streamlit>=1.33.0,<2
pandas>=2.2.0,<3
pyarrow>=15.0.0
google-cloud-bigquery>=3.20.0,<4
# Optional: enable PDF processing with Document AI
google-cloud-documentai>=2.25.0,<3
